Create spatio-time cluster.py
This is an advanced version of time series clustering. every times series has two values: 1) time invariant like features,locations 2)time variant like prices, temperature. And the time variant made up the time series.
Time series and features have different characteristics and represent different identification of a system. If we want to cluster the time series we should both measure the time series and features. This is generally called Spatio-temporal clustering method.
So in this part, we use the cluster method to forecast the future time variant in the test set.Here are the experimental steps:
Firstly, we used the spatio-temporal clustering technique similar to K-means clustering to classify different series in training set into the groups. And finding the central time series like the centroids in k-means which represent the common pattern for each group series.
Secondly, we predict the future prices belongs to which groups. Then we use the central time series as the future return to forecast. 
Finally , we use the evaluation function as our performance criterion